@@ -1,12 +1,13 @@
-name: flutter_boilerplate
+name: trifecta_app1
 description: A new Flutter project.
-
+# The following line prevents the package from being accidentally published to
+# pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 0.0.1+1
+version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
+  sdk: '>=3.1.5 <4.0.0'
 
 dependencies:
   flutter:
@@ -16,13 +17,45 @@ dependencies:
     git:
       url: git@github.com:atomic-app/atomic-sdk-flutter-releases.git
       ref: 1.0.0
-
   cupertino_icons: ^1.0.2
+  dart_jsonwebtoken: ^2.12.1
+  coast: ^2.0.2
+  flutter_riverpod: ^2.4.5
+  logger: ^2.0.2+1
+  salomon_bottom_bar: ^3.3.2
+  flutter_dotenv: ^5.1.0
+  badges: ^3.1.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^2.0.0
+  custom_lint: ^0.5.6
+  riverpod_lint: ^2.3.3
+  very_good_analysis: ^5.1.0
 
 flutter:
   uses-material-design: true
+  assets:
+    - assets/jwtRS512.key
+    - .env
+
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts from package dependencies,
+  # see https://flutter.dev/custom-fonts/#from-packages
