@@ -1,4 +1,4 @@
-package com.example.flutter_boilerplate
+package com.example.trifecta_app1
 
 //import io.flutter.embedding.android.FlutterActivity
 import io.atomic.atomic_sdk_flutter.AACFlutterActivity
